+++ sdk/ai/azure-ai-projects/tests/telemetry/test_ai_agents_instrumentor.py
@@ -624,11 +624,11 @@ class TestAiAgentsInstrumentor(AzureRecordedTestCase):
                 "attributes": {
                     "gen_ai.system": "az.ai.agents",
                     "gen_ai.thread.id": "*",
-                    # "gen_ai.agent.id": "*", - workaround for https://github.com/Azure/azure-sdk-for-python/issues/40086
+                    "gen_ai.agent.id": "*",
                     "gen_ai.thread.run.id": "*",
-                    "gen_ai.message.status": "completed",
-                    "gen_ai.usage.input_tokens": "+",
-                    "gen_ai.usage.output_tokens": "+",
+                    "gen_ai.message.status": "requires_action",
+                    # "gen_ai.usage.input_tokens": "+", # not available at the moment
+                    # "gen_ai.usage.output_tokens": "+", # not available at the moment
                     "gen_ai.event.content": '{"tool_calls": [{"id": "*", "type": "function", "function": {"name": "fetch_weather", "arguments": {"location": "New York"}}}]}',
                 },
             },
@@ -637,7 +637,7 @@ class TestAiAgentsInstrumentor(AzureRecordedTestCase):
                 "attributes": {
                     "gen_ai.system": "az.ai.agents",
                     "gen_ai.thread.id": "*",
-                    # "gen_ai.agent.id": "*", - workaround for https://github.com/Azure/azure-sdk-for-python/issues/40086
+                    "gen_ai.agent.id": "*",
                     "gen_ai.thread.run.id": "*",
                     "gen_ai.message.id": "*",
                     "gen_ai.message.status": "completed",
@@ -772,11 +772,11 @@ class TestAiAgentsInstrumentor(AzureRecordedTestCase):
                 "attributes": {
                     "gen_ai.system": "az.ai.agents",
                     "gen_ai.thread.id": "*",
-                    # "gen_ai.agent.id": "*", - workaround for https://github.com/Azure/azure-sdk-for-python/issues/40086
+                    "gen_ai.agent.id": "*",
                     "gen_ai.thread.run.id": "*",
-                    "gen_ai.message.status": "completed",
-                    "gen_ai.usage.input_tokens": "+",
-                    "gen_ai.usage.output_tokens": "+",
+                    "gen_ai.message.status": "requires_action",
+                    # "gen_ai.usage.input_tokens": "+", # not available at the moment
+                    # "gen_ai.usage.output_tokens": "+", # not available at the moment
                     "gen_ai.event.content": '{"tool_calls": [{"id": "*", "type": "function", "function": {"name": "fetch_weather", "arguments": {"location": "Sofia"}}}]}',
                 },
             },
@@ -785,7 +785,7 @@ class TestAiAgentsInstrumentor(AzureRecordedTestCase):
                 "attributes": {
                     "gen_ai.system": "az.ai.agents",
                     "gen_ai.thread.id": "*",
-                    # "gen_ai.agent.id": "*", - workaround for https://github.com/Azure/azure-sdk-for-python/issues/40086
+                    "gen_ai.agent.id": "*",
                     "gen_ai.thread.run.id": "*",
                     "gen_ai.message.id": "*",
                     "gen_ai.message.status": "completed",
@@ -953,11 +953,11 @@ class TestAiAgentsInstrumentor(AzureRecordedTestCase):
                 "attributes": {
                     "gen_ai.system": "az.ai.agents",
                     "gen_ai.thread.id": "*",
-                    # "gen_ai.agent.id": "*", - workaround for https://github.com/Azure/azure-sdk-for-python/issues/40086
+                    "gen_ai.agent.id": "*",
                     "gen_ai.thread.run.id": "*",
-                    "gen_ai.message.status": "completed",
-                    "gen_ai.usage.input_tokens": "+",
-                    "gen_ai.usage.output_tokens": "+",
+                    "gen_ai.message.status": "requires_action",
+                    # "gen_ai.usage.input_tokens": "+", # not available at the moment
+                    # "gen_ai.usage.output_tokens": "+", # not available at the moment
                     "gen_ai.event.content": '{"tool_calls": [{"id": "*", "type": "function"}]}',
                 },
             },
@@ -966,7 +966,7 @@ class TestAiAgentsInstrumentor(AzureRecordedTestCase):
                 "attributes": {
                     "gen_ai.system": "az.ai.agents",
                     "gen_ai.thread.id": "*",
-                    # "gen_ai.agent.id": "*", - workaround for https://github.com/Azure/azure-sdk-for-python/issues/40086
+                    "gen_ai.agent.id": "*",
                     "gen_ai.thread.run.id": "*",
                     "gen_ai.message.id": "*",
                     "gen_ai.message.status": "completed",