+++ sdk/ai/azure-ai-projects/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Release History
 
+## 1.0.0b8 (Unreleased)
+
+### Features added
+
+* New parameters added for Azure AI Search tool, with corresponding sample update.
+
+### Sample updates
+
+### Bugs Fixed
+
+* Fix for a bug in agent tracing causing event handler return values to not be returned when tracing is enabled.
+* Fix for a bug in agent tracing causing tool calls not to be recorded in traces.
+
+### Breaking Changes
+
 ## 1.0.0b7 (2025-03-06)
 
 ### Features added